Move workspace viewport functions back to the workspace for now
